refactor: split forgejo usecases into operation and common packages and separate tests by operation
