+++ AGENTS.md
@@ -1,4 +1,5 @@
 # AGENTS.md (docs マップ)
+Quick Decision Trees に則って、参照するべきドキュメントを全て確認すること。
 
 ## Summary
 
@@ -24,7 +25,6 @@
 - 詳細ルールをこのファイルへ増やさないこと。詳細は `docs/` 側へ追記すること。
 
 ## Quick Decision Trees
-自分がこれから行う作業を宣言しながら、作業を進めろ。
 
 ### 「タスクに着手したい」
